Bandwidth selection for the beta kernel estimator

The bandwidth is selected by minimizing the MISE using the Frank copula as the reference family. The copula parameter is set by inversion of Kendall's tau. See Nagler (2014) for details.

bw_beta(udata)

Arguments

udata
data.

Value

optimal bandwidth parameter.

Details

To speed things up, optimal bandwidths have been precalculated on a grid of tau values.

References

Nagler, T. (2014). Kernel Methods for Vine Copula Estimation. Master's Thesis, Technische Universitaet Muenchen, https://mediatum.ub.tum.de/node?id=1231221