Dependence measures of a kdecop() fit

Calculates several dependence measures derived from the copula density. All measures except "blomqvist" are computed by quasi Monte Carlo methods (see rkdecop().

dep_measures(object, measures = "all", n_qmc = 10^3, seed = 5)

Arguments

object
an object of class kdecopula.

measures
which measures to compute, see Details.

n_qmc
the number of quasi Monte Carlo samples.

seed
the seed for quasi Monte Carlo integration.

Value

A named vector of dependence measures.

The following measures are available:

"kendall"

Kendall's \(\tau\), see Nelsen (2007); computed as the sample version of a quasi Monte Carlo sample.

"spearman"

Spearman's \(\rho\), see Nelsen (2007); computed as the sample version of a quasi Monte Carlo sample.

"blomqvist"

Blomqvist's \(\beta\), see Nelsen (2007); computed as \(4C(0.5, 0.5) - 1\).

"gini"

Gini's \(\gamma\), see Nelsen (2007); computed by quasi Monte Carlo integration.

"vd_waerden"

van der Waerden's coefficient, see Genest and Verret (2005); computed as the sample version of a quasi Monte Carlo sample.

"minfo"

mutual information, see Joe (1989); computed by quasi Monte Carlo integration.

"linfoot"

Linfoot's correlation coefficient, see Joe (1989); computed by quasi Monte Carlo integration.

References

Nelsen, R. (2007). An introduction to copulas. Springer Science & Business Media, 2007.

Genest, C., and Verret, F. (2005). Locally most powerful rank tests of independence for copula models. Journal of Nonparametric Statistics, 17(5)

Joe, H. (1989). Relative Entropy Measures of Multivariate Dependence. Journal of the American Statistical Association, 84(405)

Examples

## load data and transform with empirical cdf
data(wdbc)
udat <- apply(wdbc[, -1], 2, function(x) rank(x) / (length(x) + 1))

## estimate copula density and calculate dependence measures
fit <- kdecop(udat[, 5:6])
dep_measures(fit)

#>    kendall   spearman  blomqvist       gini vd_waerden      minfo    linfoot 
#>  0.4928176  0.6814890  0.4883190  0.5505876  0.6967170  0.3153931  0.6839786